Move TestValueWithMeta and TestDumperDims to their source modules' test files
- TestValueWithMeta → test_dump_loader.py (converted from unittest to pytest)
- TestDumperDims → test_dumper.py
- Existing test_dump_loader.py tests converted from unittest to pytest style
